src/main/java/com/kpfu/javaforms/entity/AppUser.java
package com.kpfu.javaforms.entity;

import com.kpfu.javaforms.entity.base.LongEntity;
import lombok.*;
import org.hibernate.validator.constraints.Length;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

import javax.persistence.ElementCollection;
import javax.persistence.Entity;
import javax.persistence.EnumType;
import javax.persistence.Enumerated;
import javax.validation.constraints.Min;
import javax.validation.constraints.NotBlank;
import java.util.Collection;
import java.util.Set;

@Entity
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class AppUser extends LongEntity implements UserDetails {
    {
    }
    @NotBlank
    private String username;
    private String nickname;
    private String password;

    @ElementCollection
    @Enumerated(value = EnumType.STRING)
    private Set<Roles> authorities;

    @Override
    public Collection<? extends GrantedAuthority> getAuthorities() {
        return authorities;
    }

    @Override
    public String getPassword() {
        return password;
    }

    @Override
    public String getUsername() {
        return username;
    }

    @Override
    public boolean isAccountNonExpired() {
        return true;
    }

    @Override
    public boolean isAccountNonLocked() {
        return true;
    }

    @Override
    public boolean isCredentialsNonExpired() {
        return true;
    }

    @Override
    public boolean isEnabled() {
        return true;
    }

    public enum Roles implements GrantedAuthority {
        ADMIN, USER;

        @Override
        public String getAuthority() {
            return toString();
        }
    }
}
